AionTxExecSummary no longer has an RLP encoding
