+++ modMcf/src/org/aion/mcf/types/BlockIdentifier.java
@@ -1,9 +1,41 @@
 package org.aion.mcf.types;
 
-/** @author jay */
-public interface BlockIdentifier {
+import static org.aion.util.bytes.ByteUtil.byteArrayToLong;
 
-    byte[] getHash();
+import java.math.BigInteger;
+import org.aion.rlp.RLP;
+import org.aion.rlp.RLPList;
+import org.aion.util.conversions.Hex;
 
-    long getNumber();
+/** Block identifier holds block hash and number <br> */
+public class BlockIdentifier {
+
+    /** Block hash */
+    private byte[] hash;
+
+    /** Block number */
+    private long number;
+
+    public BlockIdentifier(byte[] hash, long number) {
+        this.hash = hash;
+        this.number = number;
+    }
+
+    public byte[] getHash() {
+        return hash;
+    }
+
+    public long getNumber() {
+        return number;
+    }
+
+    @Override
+    public String toString() {
+        return "BlockIdentifierImpl {"
+                + "hash="
+                + Hex.toHexString(hash)
+                + ", number="
+                + number
+                + '}';
+    }
 }
