--- modMcf/src/org/aion/mcf/types/Functional.java
@@ -1,82 +1,0 @@
-package org.aion.mcf.types;
-
-public interface Functional {
-
-    /**
-     * Represents an operation that accepts a single input argument and returns no result. Unlike
-     * most other functional interfaces, {@code Consumer} is expected to operate via side-effects.
-     *
-     * @param <T> the type of the input to the operation
-     */
-    interface Consumer<T> {
-
-        /**
-         * Performs this operation on the given argument.
-         *
-         * @param t the input argument
-         */
-        void accept(T t);
-    }
-
-    /**
-     * Represents an operation that accepts two input arguments and returns no result. This is the
-     * two-arity specialization of {@link java.util.function.Consumer}. Unlike most other functional
-     * interfaces, {@code BiConsumer} is expected to operate via side-effects.
-     *
-     * @param <T> the type of the first argument to the operation
-     * @param <U> the type of the second argument to the operation
-     * @see Functional.Consumer
-     */
-    interface BiConsumer<T, U> {
-
-        /**
-         * Performs this operation on the given arguments.
-         *
-         * @param t the first input argument
-         * @param u the second input argument
-         */
-        void accept(T t, U u);
-    }
-
-    /**
-     * Represents a function that accepts one argument and produces a result.
-     *
-     * @param <T> the type of the input to the function
-     * @param <R> the type of the result of the function
-     */
-    interface Function<T, R> {
-
-        /**
-         * Applies this function to the given argument.
-         *
-         * @param t the function argument
-         * @return the function result
-         */
-        R apply(T t);
-    }
-
-    interface Supplier<T> {
-
-        /**
-         * Gets a result.
-         *
-         * @return a result
-         */
-        T get();
-    }
-
-    interface InvokeWrapper {
-
-        void invoke();
-    }
-
-    interface InvokeWrapperWithResult<R> {
-
-        R invoke();
-    }
-
-    interface Predicate<T> {
-
-        boolean test(T t);
-    }
-}
